Refactored config POJO to remove inner classes
