+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/ConfigReader.java
@@ -1,6 +1,6 @@
 package org.ethereum.beacon.emulator.config;
 
-import org.ethereum.beacon.emulator.config.version.Config;
+import org.ethereum.beacon.emulator.config.data.Config;
 
 /** Uses object instance to construct {@link Config} */
 public interface ConfigReader {
