+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/data/v1/SignerImplementation.java
@@ -1,0 +1,24 @@
+package org.ethereum.beacon.emulator.config.data.v1;
+
+import java.util.Map;
+
+public class SignerImplementation {
+  private String clazz;
+  private Map<String, String> input;
+
+  public String getClazz() {
+    return clazz;
+  }
+
+  public void setClazz(String clazz) {
+    this.clazz = clazz;
+  }
+
+  public Map<String, String> getInput() {
+    return input;
+  }
+
+  public void setInput(Map<String, String> input) {
+    this.input = input;
+  }
+}
