+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/data/v1/Validator.java
@@ -1,0 +1,24 @@
+package org.ethereum.beacon.emulator.config.data.v1;
+
+import java.util.Map;
+
+public class Validator {
+  private Map<String, String> contract;
+  private Signer signer;
+
+  public Map<String, String> getContract() {
+    return contract;
+  }
+
+  public void setContract(Map<String, String> contract) {
+    this.contract = contract;
+  }
+
+  public Signer getSigner() {
+    return signer;
+  }
+
+  public void setSigner(Signer signer) {
+    this.signer = signer;
+  }
+}
