+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/data/v1/action/ActionEmulate.java
@@ -1,0 +1,13 @@
+package org.ethereum.beacon.emulator.config.data.v1.action;
+
+public class ActionEmulate extends Action {
+  private int count;
+
+  public int getCount() {
+    return count;
+  }
+
+  public void setCount(int count) {
+    this.count = count;
+  }
+}
