+++ emulator/config/src/main/java/org/ethereum/beacon/emulator/config/data/v1/action/ActionRun.java
@@ -1,0 +1,7 @@
+package org.ethereum.beacon.emulator.config.data.v1.action;
+
+public class ActionRun extends Action {
+  public ActionRun() {}
+
+  public ActionRun(String s) {}
+}
