+++ emulator/config/src/test/java/org/ethereum/beacon/emulator/config/ConfigReaderTest.java
@@ -1,7 +1,9 @@
 package org.ethereum.beacon.emulator.config;
 
-import org.ethereum.beacon.emulator.config.version.Config;
-import org.ethereum.beacon.emulator.config.version.ConfigV1;
+import org.ethereum.beacon.emulator.config.data.Config;
+import org.ethereum.beacon.emulator.config.data.v1.Configuration;
+import org.ethereum.beacon.emulator.config.data.v1.MainConfig;
+import org.ethereum.beacon.emulator.config.data.v1.action.ActionEmulate;
 import org.junit.Test;
 
 import java.io.File;
@@ -17,21 +19,19 @@ public class ConfigReaderTest {
     configBuilder.addYamlConfig(testYamlConfig);
     Config unmodified = configBuilder.build();
     assertEquals(1, unmodified.getVersion());
-    ConfigV1 unmodifiedV1 = (ConfigV1) unmodified;
+    MainConfig unmodifiedV1 = (MainConfig) unmodified;
     assertEquals("file://db", unmodifiedV1.getConfig().getDb());
     assertEquals("file://chainSpec.json", unmodifiedV1.getConfig().getChainSpec());
     assertEquals(3, unmodifiedV1.getPlan().getValidator().size());
     assertEquals(
-        4,
-        ((ConfigV1.PlanPart.ActionEmulate) unmodifiedV1.getPlan().getValidator().get(2))
-            .getCount());
+        4, ((ActionEmulate) unmodifiedV1.getPlan().getValidator().get(2)).getCount());
 
-    ConfigV1 config2 = new ConfigV1();
-    ConfigV1.ConfigPart configPart = config2.new ConfigPart();
+    MainConfig config2 = new MainConfig();
+    Configuration configPart = new Configuration();
     configPart.setDb("file://second/path");
     config2.setConfig(configPart);
     configBuilder.addConfig(config2);
-    ConfigV1 merged = (ConfigV1) configBuilder.build();
+    MainConfig merged = (MainConfig) configBuilder.build();
     assertEquals("file://second/path", merged.getConfig().getDb());
     assertEquals("file://chainSpec.json", merged.getConfig().getChainSpec());
     assertEquals(3, merged.getPlan().getValidator().size());
@@ -39,7 +39,7 @@ public class ConfigReaderTest {
 
     configBuilder.addConfigOverride("config.db", "file://test-db");
     configBuilder.addConfigOverride("config.validator.contract.handler", "unknown");
-    ConfigV1 overrided = (ConfigV1) configBuilder.build();
+    MainConfig overrided = (MainConfig) configBuilder.build();
     assertEquals("file://test-db", overrided.getConfig().getDb());
     assertEquals("file://chainSpec.json", overrided.getConfig().getChainSpec());
     assertEquals(3, overrided.getPlan().getValidator().size());
